+++ src/document/sgml/parser.c
@@ -28,7 +28,7 @@ static inline struct dom_node *
 add_sgml_document(struct dom_stack *stack, struct uri *uri)
 {
 	unsigned char *string = struri(uri);
-	int length = strlen(string);
+	size_t length = strlen(string);
 	struct dom_node *node = init_dom_node(DOM_NODE_DOCUMENT, string, length);
 
 	return node ? push_dom_node(stack, node) : node;
@@ -70,7 +70,7 @@ add_sgml_attribute(struct dom_stack *stack,
 	struct sgml_parser *parser = stack->parser;
 	struct dom_node *parent = get_dom_stack_top(stack)->node;
 	unsigned char *value = valtoken ? valtoken->string : NULL;
-	uint16_t valuelen = valtoken ? valtoken->length : 0;
+	size_t valuelen = valtoken ? valtoken->length : 0;
 	struct sgml_node_info *info;
 	struct dom_node *node;
 
@@ -102,11 +102,11 @@ add_sgml_proc_instruction(struct dom_stack *stack, struct scanner_token *token)
 
 	/* Anything before the separator becomes the target name ... */
 	unsigned char *name = token->string;
-	int namelen = separator ? separator - token->string : token->length;
+	size_t namelen = separator ? separator - token->string : token->length;
 
 	/* ... and everything after the instruction value. */
 	unsigned char *value = separator ? separator + 1 : NULL;
-	int valuelen = value ? token->length - namelen - 1 : 0;
+	size_t valuelen = value ? token->length - namelen - 1 : 0;
 
 	node = add_dom_proc_instruction(parent, name, namelen, value, valuelen);
 	if (!node) return NULL;
